Move test files to __tests__ directory
