+++ __tests__/controllers/postgres.test.js
@@ -1,0 +1,8 @@
+const postgres = require('../../controllers/postgres');
+const { expect } = require('chai');
+
+describe("'postgres.js' tests", () => {
+  describe('', () => {
+    it('should pend this test');
+  });
+});
